フッターおよびヘッダーコンポーネントの分離
