+++ frontend/src/components/layouts/footer/Copyright.tsx
@@ -1,0 +1,7 @@
+export const Copyright = () => {
+  return (
+    <p className="text-sm text-gray-600 dark:text-gray-400 mb-4 md:mb-0">
+      © 2024 MUSE PORT - 長崎大学マルチメディア研究会
+    </p>
+  );
+};
